requirements.txt

Flask==2.0.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
flask_marshmallow==0.14.0
marshmallow==3.13.0
